Join alpha and beta firmwares
