+++ keyboards/acheron/athena/athena.c
@@ -1,4 +1,4 @@
-/* Copyright 2020 Gondolindrim
+/* Copyright 2022 Alvaro "Gondolindrim" Volpato
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -13,33 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 #include "athena.h"
-
-
-void board_init(void) {
-    setPinInput(B6);
-    setPinInput(B7);
-}
-
-void keyboard_post_init_kb(void){
- // Defining the backlight pin (A6) as an floating (no pullup or pulldown resistor) opendrain output pin
-    palSetLineMode(BACKLIGHT_PIN, PAL_MODE_ALTERNATE(BACKLIGHT_PAL_MODE) | PAL_STM32_OTYPE_OPENDRAIN | PAL_STM32_OSPEED_HIGHEST | PAL_STM32_PUPDR_FLOATING);
-}
-
-#define LED_PIN_ON_STATE 1
-void led_init_ports(void) {
-    palSetLineMode(LED_CAPS_LOCK_PIN, PAL_MODE_OUTPUT_OPENDRAIN | PAL_STM32_PUPDR_FLOATING);
-}
-
-bool led_update_kb(led_t led_state) {
-    bool res = led_update_user(led_state);
-    #ifdef CAPSLOCK_INDICATOR
-    if(res) {
-        writePin(LED_CAPS_LOCK_PIN, !led_state.caps_lock);
-    }
-    #else
-    writePin(LED_CAPS_LOCK_PIN, 0);
-    #endif
-    return res;
-}
