+++ keyboards/acheron/athena/beta/config.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 /* USB Device descriptor parameter */
 #define VENDOR_ID 0xAC11
 #define PRODUCT_ID 0x6585 // AT+1 for Athena rev.  Beta
-#define DEVICE_VER 0x0001 
+#define DEVICE_VER 0x0002 // Rev. Beta
 #define MANUFACTURER AcheronProject
 #define PRODUCT Athena rev Beta
